requirements.txt
streamlit
pandas
gspread
google-auth
google-api-python-client
matplotlib
PyPDF2
Pillow
pytesseract
requests
pymupdf